Split JsonFormBuilder in two: JsonFormBuilder and JsonFormValueHandler.
