+++ modules/json_form_widget/src/JsonFormBuilder.php
@@ -9,7 +9,7 @@ use Drupal\metastore\SchemaRetriever;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Class.
+ * Class JsonFormBuilder.
  */
 class JsonFormBuilder implements ContainerInjectionInterface {
   use StringTranslationTrait;
@@ -309,80 +309,6 @@ class JsonFormBuilder implements ContainerInjectionInterface {
       $options = array_combine($property->enum, $property->enum);
     }
     return $options;
-  }
-
-  /**
-   * Flatten values.
-   */
-  public function flattenValues($formValues, $property, $schema) {
-    $data = [];
-
-    switch ($schema->type) {
-      case 'string':
-        $data = $this->handleStringValues($formValues, $property);
-        break;
-
-      case 'object':
-        $data = $this->handleObjectValues($formValues[$property][$property], $property, $schema);
-        break;
-
-      case 'array':
-        $data = $this->handleArrayValues($formValues, $property, $schema);
-        break;
-    }
-    return $data;
-  }
-
-  /**
-   * Flatten values for string properties.
-   */
-  public function handleStringValues($formValues, $property) {
-    if (!empty($formValues[$property])) {
-      return $formValues[$property];
-    }
-    return FALSE;
-  }
-
-  /**
-   * Flatten values for object properties.
-   */
-  public function handleObjectValues($formValues, $property, $schema) {
-    $properties = array_keys((array) $schema->properties);
-
-    $data = FALSE;
-    if (isset($formValues)) {
-      foreach ($properties as $sub_property) {
-        $value = $this->flattenValues($formValues, $sub_property, $schema->properties->$sub_property);
-        if ($value) {
-          $data[$sub_property] = $value;
-        }
-      }
-    }
-    return $data;
-  }
-
-  /**
-   * Flatten values for array properties.
-   */
-  public function handleArrayValues($formValues, $property, $schema) {
-    $data = FALSE;
-    $subschema = $schema->items;
-    if ($subschema->type === "object") {
-      foreach ($formValues[$property][$property] as $key => $item) {
-        $value = $this->handleObjectValues($formValues[$property][$property][$key][$property], $property, $schema->items);
-        if ($value) {
-          $data[$key] = $value;
-        }
-      }
-      return $data;
-    }
-
-    foreach ($formValues[$property][$property] as $key => $value) {
-      if (!empty($value)) {
-        $data[$key] = $value;
-      }
-    }
-    return $data;
   }
 
 }
